Centralize `createSingleMasterWithConnectedDeviceProtocol`.
